remove asset price in list item, if owned it already
